+++ eslint.config.mjs
@@ -1,17 +1,4 @@
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import js from "@eslint/js";
-import { FlatCompat } from "@eslint/eslintrc";
-import general from "./conf/eslint_general.mjs";
-import i18n from "./conf/eslint_i18n.mjs";
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const compat = new FlatCompat({
-  baseDirectory: __dirname,
-  recommendedConfig: js.configs.recommended,
-  allConfig: js.configs.all,
-});
+import canvaPlugin from "@canva/app-eslint-plugin";
 
 export default [
   {
@@ -23,18 +10,8 @@ export default [
       "**/*.config.*",
     ],
   },
-  ...compat.extends(
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:@typescript-eslint/eslint-recommended",
-    "plugin:@typescript-eslint/strict",
-    "plugin:@typescript-eslint/stylistic",
-    "plugin:react/recommended",
-    "plugin:jest/recommended",
-  ),
-  ...general,
-  ...i18n.map((config) => ({
-    ...config,
+  ...canvaPlugin.configs.apps_no_i18n,
+  {
     files: [
       "src/**/*",
       // Currently only the 'i18n' example is localized and following the
@@ -42,5 +19,6 @@ export default [
       // should be updated:
       "examples/i18n/**/*",
     ],
-  })),
+    ...canvaPlugin.configs.i18n,
+  },
 ];